Tool to generate graphviz data
Add a tool that generates graphviz implication graphs, like
[this](https://raw.githubusercontent.com/teorth/equational_theories/refs/heads/main/images/implications_092824.svg)

This tool needs the ability to compute graph closures and reduction, so
I've created a `graph.rb` that
transitive_closure.rb/transitive_reduction.rb/generate_graphviz_graph.rb
all use.
